data_entry/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
sqlalchemy==1.4.23
mysql-connector-python==8.0.33
databases[mysql]==0.5.3
httpx==0.23.0
python-jose[cryptography]==3.3.0
pydantic==1.8.2
cryptography==37.0.4